SYSLINK: ipc - fix heapbuf issues in open, create & delete
This patch fixes multiple issues found in the heapbuf
module code based on IPC50. The patch also includes a fix for
a minor compile warning in gatepeterson. This patch includes
the following changes.
	1. heapbuf_ioctl.h - Matched the ioctl args for open
		and create arguments. Also includes some minor
		formatting style changes.
	2. heapbuf_ioctl.c - Retrieved the correct shared
		region address pointers and gate handle from
		the user-side. Memory allocated for heapbuf name
		in heapbuf_ioctl_create and heapbuf_ioctl_open
		is adjusted as the length is already incremented
		in the user-side.
	3. heapbuf.c - gatepeterson gate is used properly
		for the gate used in internal object. The memory
		for the heapbuf object name is freed after it is
		removed from the nameserver module. This fixed
		the issue with heapbuf_delete failure. volatile
		has been used using a macro to reduce the checkpatch
		warnings.

+++ drivers/dsp/syslink/multicore_ipc/gatepeterson.c
@@ -567,7 +567,7 @@ int gatepeterson_open(void **gphandle,
 				goto noentry_fail;
 		}
 	} else
-		sharedaddr = params->shared_addr;
+		sharedaddr = (u32) params->shared_addr;
 
 	if (unlikely(((struct gatepeterson_attrs *)sharedaddr)->status !=
 						GATEPETERSON_CREATED)) {
